Example SingleCellExperiment object

This SingleCellExperiment object contains the expression values of 5 proteins (rows) from 362 cells (columns) across 3 images. The data is part of a imaging mass cytometry study on the progression of Type 1 diabetes and therefore contains pancreas cells.

pancreasSCE

Format

A SingleCellExperiment object containing the raw and arcsinh-transformed mean pixel counts per cell as well as associated cell- and protein-specific metadata. Row names represent the names of the target proteins and column names represent the image name and cell id of each cell.

colData

Cell-specific metadata where rownames represent the image name and cell id. It containins the

image number (ImageNb),

cell number/identifier (CellNb),

spatial position on the image (Pos_X, Pos_Y),

the associated image name (ImageName, see ?"pancreasImages"),

the associated mask name (MaskName, see ?"pancreasMasks"),

a arbitrary cell-type label (CellType)

a logical (Pattern) indicating exocrine cells

rowData

Protein-specific metadata where rownames represent the names of the target proteins. It contains the

channel number (frame),

metal tag of the antibody (MetalTag)

Target (the expanded name of the targeted protein)

clean_Target (the abbreviated name of the targeted protein)

assays

List of protein expression counts containing:

the raw expression counts (counts): mean pixel value per cell and protein

arcsinh-transformed raw expression counts using a co-factor of 1 (exprs)

References

Damond, N. et al., A Map of Human Type 1 Diabetes Progression by Imaging Mass Cytometry, Cell Metabolism 29:3, 2019